changed the global reference of different cards to a unique id instead of entire custom objects
